caca boudin les ptits fées (cirno)

+++ src/bullet.h
@@ -23,7 +23,7 @@ typedef struct bulletTemp {
     Vector2 location; //Corner Up-Left for drawing
     Vector2 direction;
     float speed;
-    float hp;
+    int hp;
 
     int timer;
     void (*pattern)(struct bulletTemp* bullet, float userData[MAX_DATA]);
